Add all minimum required to fully run a match
It did played a full game, by always choosing the first of any option offered.

Next steps should be to first incorporate into Evaluation the processing of
the Servers responses to Actions, since Getting a hand requires it and
right now it feels dirty to polute the game loop with that.

It also blocks me into better handling errors from the server, right now
the bot rage quits if something goes slightly wrong, as the server refusing
his play, and that forced me into bringing some of the game logic into it.

Since I plan to make something a little more challenging, the game logic
will be present, but it still could better handle non-critical errors
like errors in a running game, to loss of connection or issues with
authentication.