Refactor Haskell client for easier modularity and error handling
Not sure if the move was good, but my previous implementation was what came out in a first
try, and that didn't include ways of changing how input is collected (only bot, no human
input) or how error in plays where handled (rage quit with no LEAVE message).

The new version brings modularity, now you can instantiate the Agent which allows for
implementing an agent that collects input from user. Also, I got rid of the State Monad,
which didn't seem to fit well the architecture and that allows for easier merging of the
actuation and evaluation of game's actions (Getting hand was a pain before).

Haven't fully tested it yet, the first try could Authorize, Create, Join but then it
request to leave the game instead of deciding the Rule (it was its turn).