Add missing GAME message handler and fix issue with quotes around card play
This was trickier than I thought, as I have to convert whatever the agent receives to a
string so I can pack it and send it to server. Problem is: this is all done generically
and I needed to call show on the result to convert to a string, but showing strings cause
it to add extra quotes, the way I fixed it was by using fmap, which allows me to access
the return type inside Maybe to show it when necessary, and not doing anything on the
result of the card playing mechanic since KingCard is already a string and don't need to be
"shown".