Start a homemade State Monad for the game so that clients instances can easily get data
This is more for learning than a requirement, but I want to better handle sequencing in
clients. Things like using a do block to grab game state info where my gameloop runs the
Monad and updates with results from evaluation.

It's like the State monad on the original implementation, but I have to re-read a bunch
and have my hands dirty with some instance implementations (like Applicatives and Functor)
which was very fun.

Right now I'm missing replumbing the monad back to the evaluation function, avoiding the
pitfall of not being able to encapsulate the ZMQ context on top of my Game State.