Finish base Rust Random Bot
This is my first experience with Rust and I was able to play a full game against the bot
which was very well behaved and didn't crash or rage quit the game.

Bot currently forfeits every opportunity to Bid and always refuses offers made for his
own Positive Hands. Trump choice could be randomized but I went with Hearts always to keep
it simple.

Game Evaluation is better than in other clients, curiously, as I'm keeping a lot of info
about the game here, but I haven't added anything for the bidding, so clients don't even
know what is the current offer and who made it.

But, removing that aspect, it's very doable to create some AI for building a pruned MinMax
and trying to get some better gameplay out of the bot. I did the reading of subscription Q
less blocking by using polling, which allows process to be free for AI procesing while no
action is required from the Bot.