Refactor Haskell Client with Monads and Polling
I brought back the State Monad, it seemed like the better solution back then and I regret having
moved from it, I could solve the problem of "wrapping" the ZMQ z context within the state monad
with StateT, that solved my previous problem of dealing with ZMQ withing the State.

With that in mind I did a better architecture for the Bot, instead of typeclasses that were
mimicking the interfaces, I decided for a simpler approach, the execution is giving back to
the Main app, which executes the game loop, updating the game through a polling function that
is not blocking, allowing bot to use free time to do anything it needs.

The Bot has its own state that can be maintained, right now I'm not using it, but the idea
is that I can store calculations done while idling and use them to make an informed play later.
Access to the game is in a read only form, through the Reader Monad, all "wrapping" the ZMQ context,
which allows bot to take whatever actions needed on the Server.

Tested running a full game against Rust Bot, Python Bot and User playing on HTML Client. Bot
performance was really fast, strategy is very weak though (firstValid).